Lab done

+++ src/main/java/beans/Entry.java
@@ -1,12 +1,12 @@
 package beans;
 
 public class Entry {
-    private float x;
-    private int y;
+    private double x;
+    private double y;
     private int r;
     private boolean hitResult;
 
-    public Entry(float x, int y, int r) {
+    public Entry(double x, double y, int r) {
         this.x = x;
         this.y = y;
         this.r = r;
@@ -20,19 +20,19 @@ public class Entry {
         this.hitResult = hitResult;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
